+++ src/PVOutput.Net/Modules/SystemService.cs
@@ -61,10 +61,10 @@ namespace PVOutput.Net.Modules
         /// </summary>
         /// <param name="systemId">The system to update.</param>
         /// <param name="newSystemName">New system name.</param>
-        /// <param name="configurations">List of modified extended data definitions.</param>
+        /// <param name="dataDefinitions">List of modified extended data definitions.</param>
         /// <param name="cancellationToken">A cancellation token for the request.</param>
         /// <returns>If the operation succeeded.</returns>
-        public Task<PVOutputBasicResponse> PostSystem(int systemId, string newSystemName = null, IEnumerable<IExtendedDataDefinition> configurations = null, CancellationToken cancellationToken = default)
+        public Task<PVOutputBasicResponse> PostSystem(int systemId, string newSystemName = null, IEnumerable<IExtendedDataDefinition> dataDefinitions = null, CancellationToken cancellationToken = default)
         {
             var loggingScope = new Dictionary<string, object>()
             {
@@ -73,7 +73,7 @@ namespace PVOutput.Net.Modules
             };
 
             var handler = new RequestHandler(Client);
-            return handler.ExecutePostRequestAsync(new PostSystemRequest() { SystemId = systemId, SystemName = newSystemName, Configurations = configurations }, loggingScope, cancellationToken);
+            return handler.ExecutePostRequestAsync(new PostSystemRequest() { SystemId = systemId, SystemName = newSystemName, DataDefinitions = dataDefinitions }, loggingScope, cancellationToken);
         }
     }
 }